third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-createOffer_rest-expected.txt

This is a testharness.js-based test.
[FAIL] When media stream is added when createOffer() is running in parallel, the result offer should contain the new media stream
  assert_equals: Expect m=audio line to be found in offer SDP expected 1 but got 0
Harness: the test ran to completion.
